+++ src/smartcontracts/smartContract.local.net.spec.ts
@@ -15,6 +15,7 @@ import { ResultsParser } from "./resultsParser";
 
 describe("test on local testnet", function () {
     let provider = chooseProxyProvider("local-testnet");
+    let watcher = new TransactionWatcher(provider);
     let alice: TestWallet, bob: TestWallet, carol: TestWallet;
     let resultsParser = new ResultsParser();
 
@@ -28,14 +29,15 @@ describe("test on local testnet", function () {
         TransactionWatcher.DefaultPollingInterval = 5000;
         TransactionWatcher.DefaultTimeout = 50000;
 
-        await NetworkConfig.getDefault().sync(provider);
+        let network = await provider.getNetworkConfig();
         await alice.sync(provider);
 
         // Deploy
         let contract = new SmartContract({});
         let transactionDeploy = contract.deploy({
             code: await loadContractCode("src/testdata/counter.wasm"),
-            gasLimit: new GasLimit(3000000)
+            gasLimit: new GasLimit(3000000),
+            chainID: network.ChainID
         });
 
         transactionDeploy.setNonce(alice.account.nonce);
@@ -46,7 +48,8 @@ describe("test on local testnet", function () {
         // ++
         let transactionIncrement = contract.call({
             func: new ContractFunction("increment"),
-            gasLimit: new GasLimit(3000000)
+            gasLimit: new GasLimit(3000000),
+            chainID: network.ChainID
         });
 
         transactionIncrement.setNonce(alice.account.nonce);
@@ -57,12 +60,14 @@ describe("test on local testnet", function () {
         // Now, let's build a few transactions, to be simulated
         let simulateOne = contract.call({
             func: new ContractFunction("increment"),
-            gasLimit: new GasLimit(100000)
+            gasLimit: new GasLimit(100000),
+            chainID: network.ChainID
         });
 
         let simulateTwo = contract.call({
             func: new ContractFunction("foobar"),
-            gasLimit: new GasLimit(500000)
+            gasLimit: new GasLimit(500000),
+            chainID: network.ChainID
         });
 
         simulateOne.setNonce(alice.account.nonce);
@@ -75,12 +80,12 @@ describe("test on local testnet", function () {
         await transactionDeploy.send(provider);
         await transactionIncrement.send(provider);
 
-        await transactionDeploy.awaitExecuted(provider);
+        await watcher.awaitCompleted(transactionDeploy);
         let transactionOnNetwork = await transactionDeploy.getAsOnNetwork(provider);
         let bundle = resultsParser.parseUntypedOutcome(transactionOnNetwork);
         assert.isTrue(bundle.returnCode.isSuccess());
 
-        await transactionIncrement.awaitExecuted(provider);
+        await watcher.awaitCompleted(transactionIncrement);
         transactionOnNetwork = await transactionDeploy.getAsOnNetwork(provider);
         bundle = resultsParser.parseUntypedOutcome(transactionOnNetwork);
         assert.isTrue(bundle.returnCode.isSuccess());
@@ -96,14 +101,15 @@ describe("test on local testnet", function () {
         TransactionWatcher.DefaultPollingInterval = 5000;
         TransactionWatcher.DefaultTimeout = 50000;
 
-        await NetworkConfig.getDefault().sync(provider);
+        let network = await provider.getNetworkConfig();
         await alice.sync(provider);
 
         // Deploy
         let contract = new SmartContract({});
         let transactionDeploy = contract.deploy({
             code: await loadContractCode("src/testdata/counter.wasm"),
-            gasLimit: new GasLimit(3000000)
+            gasLimit: new GasLimit(3000000),
+            chainID: network.ChainID
         });
 
         transactionDeploy.setNonce(alice.account.nonce);
@@ -114,7 +120,8 @@ describe("test on local testnet", function () {
         // ++
         let transactionIncrementFirst = contract.call({
             func: new ContractFunction("increment"),
-            gasLimit: new GasLimit(2000000)
+            gasLimit: new GasLimit(2000000),
+            chainID: network.ChainID
         });
 
         transactionIncrementFirst.setNonce(alice.account.nonce);
@@ -125,7 +132,8 @@ describe("test on local testnet", function () {
         // ++
         let transactionIncrementSecond = contract.call({
             func: new ContractFunction("increment"),
-            gasLimit: new GasLimit(2000000)
+            gasLimit: new GasLimit(2000000),
+            chainID: network.ChainID
         });
 
         transactionIncrementSecond.setNonce(alice.account.nonce);
@@ -138,9 +146,9 @@ describe("test on local testnet", function () {
         await transactionIncrementFirst.send(provider);
         await transactionIncrementSecond.send(provider);
 
-        await transactionDeploy.awaitExecuted(provider);
-        await transactionIncrementFirst.awaitExecuted(provider);
-        await transactionIncrementSecond.awaitExecuted(provider);
+        await watcher.awaitCompleted(transactionDeploy);
+        await watcher.awaitCompleted(transactionIncrementFirst);
+        await watcher.awaitCompleted(transactionIncrementSecond);
 
         // Check counter
         let queryResponse = await contract.runQuery(provider, { func: new ContractFunction("get") });
@@ -153,7 +161,7 @@ describe("test on local testnet", function () {
         TransactionWatcher.DefaultPollingInterval = 5000;
         TransactionWatcher.DefaultTimeout = 50000;
 
-        await NetworkConfig.getDefault().sync(provider);
+        let network = await provider.getNetworkConfig();
         await alice.sync(provider);
 
         // Deploy
@@ -161,7 +169,8 @@ describe("test on local testnet", function () {
         let transactionDeploy = contract.deploy({
             code: await loadContractCode("src/testdata/erc20.wasm"),
             gasLimit: new GasLimit(50000000),
-            initArguments: [new U32Value(10000)]
+            initArguments: [new U32Value(10000)],
+            chainID: network.ChainID
         });
 
         // The deploy transaction should be signed, so that the address of the contract
@@ -174,13 +183,15 @@ describe("test on local testnet", function () {
         let transactionMintBob = contract.call({
             func: new ContractFunction("transferToken"),
             gasLimit: new GasLimit(9000000),
-            args: [new AddressValue(bob.address), new U32Value(1000)]
+            args: [new AddressValue(bob.address), new U32Value(1000)],
+            chainID: network.ChainID
         });
 
         let transactionMintCarol = contract.call({
             func: new ContractFunction("transferToken"),
             gasLimit: new GasLimit(9000000),
-            args: [new AddressValue(carol.address), new U32Value(1500)]
+            args: [new AddressValue(carol.address), new U32Value(1500)],
+            chainID: network.ChainID
         });
 
         // Apply nonces and sign the remaining transactions
@@ -197,9 +208,9 @@ describe("test on local testnet", function () {
         await transactionMintBob.send(provider);
         await transactionMintCarol.send(provider);
 
-        await transactionDeploy.awaitExecuted(provider);
-        await transactionMintBob.awaitExecuted(provider);
-        await transactionMintCarol.awaitExecuted(provider);
+        await watcher.awaitCompleted(transactionDeploy);
+        await watcher.awaitCompleted(transactionMintBob);
+        await watcher.awaitCompleted(transactionMintCarol);
 
         // Query state, do some assertions
         let queryResponse = await contract.runQuery(provider, {
@@ -232,7 +243,7 @@ describe("test on local testnet", function () {
         TransactionWatcher.DefaultPollingInterval = 5000;
         TransactionWatcher.DefaultTimeout = 50000;
 
-        await NetworkConfig.getDefault().sync(provider);
+        let network = await provider.getNetworkConfig();
         await alice.sync(provider);
 
         // Deploy
@@ -240,7 +251,8 @@ describe("test on local testnet", function () {
         let transactionDeploy = contract.deploy({
             code: await loadContractCode("src/testdata/lottery-esdt.wasm"),
             gasLimit: new GasLimit(50000000),
-            initArguments: []
+            initArguments: [],
+            chainID: network.ChainID
         });
 
         // The deploy transaction should be signed, so that the address of the contract
@@ -263,7 +275,8 @@ describe("test on local testnet", function () {
                 OptionValue.newMissing(),
                 OptionValue.newMissing(),
                 OptionalValue.newMissing()
-            ]
+            ],
+            chainID: network.ChainID
         });
 
         // Apply nonces and sign the remaining transactions
@@ -275,8 +288,8 @@ describe("test on local testnet", function () {
         await transactionDeploy.send(provider);
         await transactionStart.send(provider);
 
-        await transactionDeploy.awaitNotarized(provider);
-        await transactionStart.awaitNotarized(provider);
+        await watcher.awaitAllEvents(transactionDeploy, ["SCDeploy"]);
+        await watcher.awaitAnyEvent(transactionStart, ["completedTxEvent"]);
 
         // Let's check the SCRs
         let transactionOnNetwork = await transactionDeploy.getAsOnNetwork(provider);