+++ src/smartcontracts/smartContractResults.local.net.spec.ts
@@ -9,7 +9,8 @@ import { ContractFunction } from "./function";
 import { ResultsParser } from "./resultsParser";
 
 describe("fetch transactions from local testnet", function () {
-    let provider = chooseProxyProvider("local-testnet");;
+    let provider = chooseProxyProvider("local-testnet");
+    let watcher = new TransactionWatcher(provider);
     let alice: TestWallet;
     let resultsParser = new ResultsParser();
 
@@ -23,14 +24,15 @@ describe("fetch transactions from local testnet", function () {
         TransactionWatcher.DefaultPollingInterval = 5000;
         TransactionWatcher.DefaultTimeout = 50000;
 
-        await NetworkConfig.getDefault().sync(provider);
+        let network = await provider.getNetworkConfig();
         await alice.sync(provider);
 
         // Deploy
         let contract = new SmartContract({});
         let transactionDeploy = contract.deploy({
             code: await loadContractCode("src/testdata/counter.wasm"),
-            gasLimit: new GasLimit(3000000)
+            gasLimit: new GasLimit(3000000),
+            chainID: network.ChainID
         });
 
         transactionDeploy.setNonce(alice.account.nonce);
@@ -41,7 +43,8 @@ describe("fetch transactions from local testnet", function () {
         // ++
         let transactionIncrement = contract.call({
             func: new ContractFunction("increment"),
-            gasLimit: new GasLimit(3000000)
+            gasLimit: new GasLimit(3000000),
+            chainID: network.ChainID
         });
 
         transactionIncrement.setNonce(alice.account.nonce);
@@ -53,8 +56,8 @@ describe("fetch transactions from local testnet", function () {
         await transactionDeploy.send(provider);
         await transactionIncrement.send(provider);
 
-        await transactionDeploy.awaitExecuted(provider);
-        await transactionIncrement.awaitExecuted(provider);
+        await watcher.awaitCompleted(transactionDeploy);
+        await watcher.awaitCompleted(transactionIncrement);
 
         await transactionDeploy.getAsOnNetwork(provider);
         await transactionIncrement.getAsOnNetwork(provider);
@@ -66,15 +69,5 @@ describe("fetch transactions from local testnet", function () {
         transactionOnNetwork = transactionIncrement.getAsOnNetworkCached();
         bundle = resultsParser.parseUntypedOutcome(transactionOnNetwork);
         assert.isTrue(bundle.returnCode.isSuccess());
-    });
-
-    it("ESDT", async function () {
-        this.timeout(60000);
-
-        TransactionWatcher.DefaultPollingInterval = 5000;
-        TransactionWatcher.DefaultTimeout = 50000;
-
-        await NetworkConfig.getDefault().sync(provider);
-        await alice.sync(provider);
     });
 });